Implement AVX Optimization
I rewrite almost all the funtions in `math-toolkit.h` with AVX intrinsics
in the hope of increasing the program's performance. Sadly, after I
have tried many methods, the outcome shows that the more AVX intrinsics
or SIMD instructions involve in the functions, the longer the execution
time will be take by the program.